quick fix

+++ bit_manipulation/3-set_bit.c
@@ -9,7 +9,7 @@
  * Return: 1 on success, -1 on fail
  */
 
-int set_bit(unsigned long int *n, unsigned long int index)
+int set_bit(unsigned long int *n, unsigned int index)
 {
 	unsigned int i = 1;
